Do not overwrite config-default.conf
Previously, when config-example.conf was missing, it would be created
and config-default.conf would be made a symlink to it. However, this
would overwrite any existing config-default.conf, potentially throwing
away a user's data. This occurs when you clean up and throw away the
(cluttering) config-example.conf (and depending on whether you use a
symlink to another named config, or put data in config-default.conf,
this would either just replace the symlink, or throw away your actual
config data).

This commit splits the update checks, so no files are ever overwritten,
except a broken config-default.conf symlink.

+++ compile.sh
@@ -224,6 +224,10 @@ if ! ls "${SRC}"/userpatches/{config-example.conf,config-docker.conf,config-vagr
 	# Create example config
 	if [[ ! -f "${SRC}"/userpatches/config-example.conf ]]; then
 		cp "${SRC}"/config/templates/config-example.conf "${SRC}"/userpatches/config-example.conf || exit 1
+	fi
+
+	# Link default config to example config
+	if [[ ! -f "${SRC}"/userpatches/config-default.conf ]]; then
                 ln -fs config-example.conf "${SRC}"/userpatches/config-default.conf || exit 1
 	fi
